requirements.txt
streamlit
torch
transformers
sentence-transformers
faiss-cpu
pdfplumber
numpy
gtts
SpeechRecognition
pyaudio